# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creatrid API
  description: |
    Creatrid is a verified digital identity platform for creators. Connect social
    accounts, build a Creator Score, manage content licensing, and share a public
    profile.

    ## Authentication
    Most endpoints require authentication via an HTTP-only cookie named `token`
    that contains a JWT issued after Google OAuth sign-in.

    Third-party verification endpoints use API key authentication via the
    `X-API-Key` header.
  version: 1.0.0
  contact:
    name: Creatrid
    url: https://creatrid.com
servers:
  - url: https://creatrid-api.bluewave-351c9003.westeurope.azurecontainerapps.io
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/auth/google/callback:
    get:
      operationId: googleCallback
      tags: [Auth]
      summary: Google OAuth callback
      description: |
        Handles the OAuth callback from Google. Validates state, exchanges the
        authorization code for tokens, finds or creates the user, issues a JWT
        cookie, and redirects to the frontend.
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Authorization code from Google
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: CSRF state parameter
      responses:
        "307":
          description: Redirect to frontend (dashboard or onboarding)
  /api/auth/me:
    get:
      operationId: getCurrentUser
      tags: [Auth]
      summary: Get current user
      description: Returns the currently authenticated user's full profile.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Log out
      description: Clears the JWT cookie to end the session.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Successfully logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/verify-email/send:
    post:
      operationId: sendEmailVerification
      tags: [Auth]
      summary: Send email verification
      description: Sends a verification email to the authenticated user's email address.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Email already verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/verify-email/{token}:
    get:
      operationId: verifyEmail
      tags: [Auth]
      summary: Verify email address
      description: Verifies the user's email address using the token from the verification email. Redirects to the frontend.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
          description: Email verification token
      responses:
        "302":
          description: Redirect to frontend with verification status

  # ──────────────────────────────────────────────
  # Users
  # ──────────────────────────────────────────────
  /api/users/onboard:
    post:
      operationId: onboardUser
      tags: [Users]
      summary: Complete onboarding
      description: Sets the user's username and display name during first-time onboarding.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, name]
              properties:
                username:
                  type: string
                  pattern: "^[a-zA-Z0-9_-]{3,30}$"
                  description: Unique username (3-30 chars, letters, numbers, hyphens, underscores)
                  example: johndoe
                name:
                  type: string
                  description: Display name
                  example: John Doe
      responses:
        "200":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error (invalid username, taken, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/profile:
    patch:
      operationId: updateProfile
      tags: [Users]
      summary: Update profile
      description: Updates the authenticated user's profile fields. All fields are optional.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display name
                bio:
                  type: string
                  description: Bio text
                username:
                  type: string
                  pattern: "^[a-zA-Z0-9_-]{3,30}$"
                  description: New username
                theme:
                  type: string
                  enum: [default, ocean, sunset, forest, midnight, rose]
                  description: Profile theme
                customLinks:
                  type: array
                  maxItems: 10
                  items:
                    $ref: "#/components/schemas/CustomLink"
                  description: Custom links (max 10)
                emailPrefs:
                  $ref: "#/components/schemas/EmailPrefs"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/profile/image:
    post:
      operationId: uploadProfileImage
      tags: [Users]
      summary: Upload profile image
      description: Uploads a new profile photo. Accepts JPEG, PNG, and WebP up to 5 MB. Images are resized to 512x512.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, or WebP, max 5 MB)
      responses:
        "200":
          description: Image uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  image:
                    type: string
                    format: uri
                    description: URL of the uploaded image
        "400":
          description: Invalid file or file too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          description: Image upload not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/account:
    delete:
      operationId: deleteAccount
      tags: [Users]
      summary: Delete account
      description: Permanently deletes the authenticated user's account and all associated data.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                    example: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/export:
    get:
      operationId: exportProfile
      tags: [Users]
      summary: Export profile data
      description: Downloads the authenticated user's profile data as a JSON file (GDPR data export).
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Profile data export
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/PublicUser"
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicConnection"
                  exportedAt:
                    type: string
                    format: date-time
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/users/{username}:
    get:
      operationId: getPublicProfile
      tags: [Users]
      summary: Get public profile
      description: Returns the public profile for a creator by username.
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Public profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/PublicUser"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/users/{username}/connections:
    get:
      operationId: getPublicConnections
      tags: [Users]
      summary: Get public connections
      description: Returns the social platform connections for a creator by username.
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: Public connections
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicConnection"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/users/{username}/content:
    get:
      operationId: getPublicContent
      tags: [Users]
      summary: Get public content
      description: Returns the public content items for a creator by username.
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [image, video, audio, text, other]
          description: Filter by content type
        - name: q
          in: query
          schema:
            type: string
          description: Search query for title or description
      responses:
        "200":
          description: Public content list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ContentItemPublic"
                  total:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"
  /api/users/{username}/collections:
    get:
      operationId: getPublicCollections
      tags: [Users]
      summary: Get public collections
      description: Returns the public collections for a creator by username.
      parameters:
        - $ref: "#/components/parameters/Username"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Public collections
          content:
            application/json:
              schema:
                type: object
                properties:
                  collections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Collection"
                  total:
                    type: integer
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # Connections
  # ──────────────────────────────────────────────
  /api/connections:
    get:
      operationId: listConnections
      tags: [Connections]
      summary: List connections
      description: Returns the authenticated user's social platform connections.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Connections list
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicConnection"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/connections/{platform}/connect:
    get:
      operationId: connectPlatform
      tags: [Connections]
      summary: Start platform connection
      description: |
        Redirects the user to the platform's OAuth consent screen. Requires browser-based
        authentication (redirect auth middleware).
      parameters:
        - $ref: "#/components/parameters/Platform"
      security:
        - cookieAuth: []
      responses:
        "307":
          description: Redirect to platform OAuth
  /api/connections/{platform}/callback:
    get:
      operationId: platformCallback
      tags: [Connections]
      summary: Platform OAuth callback
      description: Handles the OAuth callback from the social platform. Exchanges the code, fetches the profile, and stores the connection.
      parameters:
        - $ref: "#/components/parameters/Platform"
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
      security:
        - cookieAuth: []
      responses:
        "307":
          description: Redirect to frontend connections page
  /api/connections/{platform}:
    delete:
      operationId: disconnectPlatform
      tags: [Connections]
      summary: Disconnect platform
      description: Removes the connection for the specified platform.
      parameters:
        - $ref: "#/components/parameters/Platform"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Platform disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/connections/{platform}/refresh:
    post:
      operationId: refreshConnection
      tags: [Connections]
      summary: Refresh connection
      description: Refreshes the token and re-fetches the profile data for a platform connection.
      parameters:
        - $ref: "#/components/parameters/Platform"
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Connection refresh result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [refreshed, refresh_failed, fetch_failed, no_token]
                    description: Result of the refresh attempt
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # Analytics
  # ──────────────────────────────────────────────
  /api/analytics:
    get:
      operationId: getAnalyticsSummary
      tags: [Analytics]
      summary: Get analytics summary
      description: Returns profile view and click analytics for the authenticated user.
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token set as an HTTP-only cookie after Google OAuth sign-in
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party verification endpoints (starts with crk_)
...
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: Not authenticated
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    CustomLink:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
...